Added mule yaml file to test s3 functions.
